colour shader
Entering game mode will currently result in an error message reading "PropertyToID can only be called from the main thread". This is due to the call to UpdateMeshHeights from the MapGenerator, which, when in game mode, is running on a separate thread (as part of our optimization efforts). Will fix later